refactored repository extensions to work with persisted repository for issue #25
